refactor: move Topbar into Layout
