+++ src/common/components/account-modal/account-modal.tsx
@@ -1,20 +1,32 @@
+
 import React, { ReactElement } from 'react';
-import { Button, Modal } from 'react-bootstrap';
-import { useDispatch, useSelector, useStore } from 'react-redux';
+import {
+  Button,
+  Modal
+} from 'react-bootstrap';
+import {
+  useDispatch,
+  useSelector,
+  useStore
+} from 'react-redux';
+import { useTranslation } from 'react-i18next';
+
 import { StoreType } from '../../types/util.types';
 import { showAccountModalAction } from '../../actions/general.actions';
-import { useTranslation } from 'react-i18next';
 import fetchIssueTransactions from '../../live-data/issue-transaction.watcher';
-
 import './account-modal.scss';
 
 type AccountModalProps = {
-    onSelected: (account: string) => void | Promise<void>;
-    selected?: string;
+  onSelected: (account: string) => void | Promise<void>;
+  selected?: string;
 };
 
-export default function AccountModal(props: AccountModalProps): ReactElement {
-  const { showAccountModal, accounts, extensions } = useSelector((state: StoreType) => state.general);
+function AccountModal(props: AccountModalProps): ReactElement {
+  const {
+    showAccountModal,
+    accounts,
+    extensions
+  } = useSelector((state: StoreType) => state.general);
   const store = useStore();
   const { t } = useTranslation();
   const dispatch = useDispatch();
@@ -33,7 +45,7 @@ export default function AccountModal(props: AccountModalProps): ReactElement {
       </Modal.Header>
       <Modal.Body className='account-modal'>
         {extensions.length ? (
-          <React.Fragment>
+          <>
             {!accounts.length && (
               <div id='account-modal-no-account'>
                 <p>
@@ -42,13 +54,13 @@ export default function AccountModal(props: AccountModalProps): ReactElement {
                     href='https://polkadot.js.org/extension/'
                     target='_blank'
                     rel='noopener noreferrer'>
-                                        &nbsp;{t('here')}
+                    &nbsp;{t('here')}
                   </a>
-                                    . {t('refresh_page')}
+                  . {t('refresh_page')}
                 </p>
               </div>
             )}
-            {(accounts || []).map((account: string, index: number) => (
+            {accounts?.map((account: string, index: number) => (
               <div
                 className='row'
                 key={index}>
@@ -64,9 +76,9 @@ export default function AccountModal(props: AccountModalProps): ReactElement {
                 </div>
               </div>
             ))}
-          </React.Fragment>
+          </>
         ) : (
-          <React.Fragment>
+          <>
             <div className='row description'>
               <div className='col-12'>{t('install_supported_wallets')}</div>
             </div>
@@ -87,16 +99,18 @@ export default function AccountModal(props: AccountModalProps): ReactElement {
                 </a>
               </div>
             </div>
-          </React.Fragment>
+          </>
         )}
       </Modal.Body>
       <Modal.Footer>
         <Button
           variant='secondary'
           onClick={onClose}>
-                    Close
+          Close
         </Button>
       </Modal.Footer>
     </Modal>
   );
 }
+
+export default AccountModal;
